+++ src/ray.rs
@@ -8,5 +8,5 @@ pub struct Ray {
 
 impl Ray {
     pub fn new(origin: Point3, direction: Vec3) -> Self { Self { origin, direction } }
-    pub fn at(self, t: f64) -> Point3 { self.origin + t * self.direction }
+    pub fn at(self, t: f32) -> Point3 { self.origin + t * self.direction }
 }
