+++ src/texture.rs
@@ -3,11 +3,11 @@ use crate::vec3::{Color, Point3};
 #[derive(Clone, Copy)]
 pub enum Texture {
     Solid(Color),
-    Checker { scale: f64, even: Color, odd: Color },
+    Checker { scale: f32, even: Color, odd: Color },
 }
 
 impl Texture {
-    pub fn value(&self, _u: f64, _v: f64, p: Point3) -> Color {
+    pub fn value(&self, _u: f32, _v: f32, p: Point3) -> Color {
         match self {
             Texture::Solid(c) => *c,
             Texture::Checker { scale, even, odd } => {
